done car company, vehicle, seat, ticket, coupon api need test

+++ src/main/java/com/be_planfortrips/repositories/PaymentRepository.java
@@ -3,5 +3,5 @@ package com.be_planfortrips.repositories;
 import com.be_planfortrips.entity.Payment;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PaymentRepository extends JpaRepository<Payment, Long> {
+public interface PaymentRepository extends JpaRepository<Payment, Integer> {
 }
